upgrade to err-backend-discord 4.0.0
