+++ src/errbot/backend/err-backend-discord/requirements.txt
@@ -1,1 +1,1 @@
-discord.py==1.6.0
+discord.py==2.0.1
